bit shift bug fix after division

+++ sensor.py
@@ -216,7 +216,7 @@ class Sensor_BM280:
         if (var1 == 0):
             return 0 #// avoid exception caused by division by zero
         p = 1048576-adc_P
-        p = (((p<<31)-var2)*3125)/var1
+        p = int((((p<<31)-var2)*3125)/var1)
         var1 = ((dig_P9) * (p>>13) * (p>>13)) >> 25
         var2 = ((dig_P8) * p) >> 19
         p = ((p + var1 + var2) >> 8) + ((dig_P7)<<4)
